bugfix: 라우팅 관련 오류 수정 (#42)
