+++ src/app/(mypage)/account/page.tsx
