+++ src/components/features/nav-mypage/mypage-button.tsx
@@ -50,7 +50,7 @@ const MypageButton = () => {
       router.push('/sign-in');
       return;
     }
-    router.push(`/mypage/${path}`);
+    router.push(`/${path}`);
     setClose();
   };
